chore(common): move `MouseContext` to sub module for symmetry reasons
