+++ src/plugins/menu/src/systems/update_panels/activity_colors_override.rs
@@ -16,7 +16,7 @@ use bevy::{
 	ui::BackgroundColor,
 };
 use common::{
-	states::MouseContext,
+	states::mouse_context::MouseContext,
 	traits::{accessors::get::GetterRef, iterate::Iterate, map_value::TryMapBackwards},
 };
 use player::components::player::Player;
