+++ src/plugins/skills/src/systems/mouse_context/release.rs
@@ -3,7 +3,7 @@ use bevy::{
 	input::{mouse::MouseButton, ButtonInput},
 	state::state::{NextState, State},
 };
-use common::states::MouseContext;
+use common::states::mouse_context::MouseContext;
 
 pub(crate) fn release_triggered_mouse_context(
 	mouse_input: Res<ButtonInput<MouseButton>>,
